dcp-t4(001): move cobracompletion into factory_definitions
Relocate pkg/transports/cli/cobracompletion to
pkg/services/factory_definitions/transports/cli/cobracompletion. The package
projects Factory Definitions catalog state onto Cobra completion, so it
belongs in that service's own transports subtree per the packaged-structure
convention.

The relocation is a git mv with unchanged package names and test contents.
Six external importers plus the three self-imports in the package's own
external tests are retargeted.

One production import could not move as-is. Since the PRD for this lane was
written, #2057 moved climanifest into the Work service transports tree, so
selected_factory_signature.go's use of climanifest.ComposeRunInputs became a
Factory-Definitions-to-Work peer import the moment the package landed under
pkg/services. cmd/pkgboundarycheck rejects that edge: main.go:1871
isMatchingServiceOwnedTransportConsumer only excepts consumers under the
matching composition-root protocol root (pkg/transports/cli), never a peer
service's transport subtree.

That file now consumes the composition root instead, through a new thin
delegating package pkg/transports/cli/climanifestentry. This is the pattern
exact case: "Composition-root transports and peer-service transports consume
this package instead of reaching into that service's transport subpackages
directly." Manifest and EffectiveInputSchema are type aliases, so every call
site keeps passing and receiving the identical climanifest types and no
caller signature changes.

Behavior is unchanged. cmd/servicecyclecheck still measures minimum feedback
arc weight 22, matching the ceiling, and pkg-boundary passes with no new
baseline entry.

The package's test-only import of pkg/services/work is deliberately left in
place and unmodified, not hidden or restructured; cmd/servicecyclecheck skips
_test.go files by construction (graph.go:134), so it carries no cross-service
weight.
