+++ pkg/services/factory_definitions/transports/cli/cobracompletion/factory_names.go
