+++ pkg/services/factory_definitions/transports/cli/cobracompletion/packaged_factory_names.go
