+++ pkg/services/factory_definitions/transports/cli/cobracompletion/powershell_test.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/portpowered/infinite-you/pkg/transports/cli/cobracompletion"
+	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/transports/cli/cobracompletion"
 	"github.com/spf13/cobra"
 )
