+++ pkg/services/factory_definitions/transports/cli/cobracompletion/selected_factory_signature_test.go
@@ -6,9 +6,9 @@ import (
 	"testing"
 
 	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
+	"github.com/portpowered/infinite-you/pkg/services/factory_definitions/transports/cli/cobracompletion"
 	"github.com/portpowered/infinite-you/pkg/services/work"
 	"github.com/portpowered/infinite-you/pkg/services/work/transports/cli/climanifest"
-	"github.com/portpowered/infinite-you/pkg/transports/cli/cobracompletion"
 	"github.com/spf13/cobra"
 )
